[review] Move eth rpcs dependend on GM to separate module
